+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.5.0] - 2022-03-26
+### Added
+- AddComponentMenu ordering
+- MenuItem intergration
+### Fixed
+- Menu Editor Visualizer organize bug
+
 ## [1.4.2] - 2022-02-22
 ### Fixed
 - Bug in MenuEditorVisualizer